Refactor out DrawPath
